Changesd a couple file names and moved files reassured the imports linked properly.
