+++ client/src/main.jsx
@@ -3,11 +3,11 @@ import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 
 import App from './App.jsx';
 import Home from './pages/Home/Home.jsx';
-import About from './pages/About.jsx';
-import Products from './pages/Products.jsx';
-import Reviews from './pages/Reviews.jsx';
-import Why from './pages/Why.jsx';
-import Login from './pages/Login.jsx';
+import About from './pages/About/About.jsx';
+import Products from './pages/Products/Products.jsx';
+import Reviews from './pages/Reviews/Reviews.jsx';
+import Why from './pages/Why/Why.jsx';
+import Login from './pages/Login/Login.jsx';
 import './index.css';
 
 const router = createBrowserRouter([
